requirements.txt
alembic==1.7.7
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Login==0.6.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
SQLAlchemy==1.4.34
Werkzeug==2.1.1
WTForms==3.0.1
